Start Screen
The game now has a start screen! When you click start you are given the option to select your class and when you do that game begins. This allows me to start making other classes, and they should be as easy to implement as the abilities in the ability system.